+++ HolaSDL/Animation.h
@@ -19,7 +19,7 @@ public:
 	void runAnimation();
 	void startAnimation();
 	bool isFinished();
-	void setTime(int tim);
+	void setTime(float tim);
 	void addAnimationFrame(SDL_Rect* srcRect, SDL_Rect destRect, int xOffset=0, int yOffset=0, int frameWidth=64, int frameHeight=64);
 	void setLayer(TileLayer* lay);
 	TileLayer* getLayer() { return layer; };
